Add parameter module inst for native wrapper functions (#117)
And add asm code of em64/arm/mips version to call native wrapper functions;
Fix some issues of calling wrapper functions;

+++ core/iwasm/lib/app-libs/extension/gui/src/wgl_label.c
@@ -16,12 +16,12 @@
 
 
 #include "wgl.h"
-#include "native_interface.h"
+#include "gui_api.h"
 #include <string.h>
 
 
 #define ARGC sizeof(argv)/sizeof(uint32)
-#define CALL_LABEL_NATIVE_FUNC(id) wasm_label_native_call(id, (uint32)argv, ARGC)
+#define CALL_LABEL_NATIVE_FUNC(id) wasm_label_native_call(id, argv, ARGC)
 
 wgl_obj_t wgl_label_create(wgl_obj_t par, wgl_obj_t copy)
 {
